# internal/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API request bodies and responses are usually JSON-encoded, except for binary data and some of the OAuth2 endpoints. Note that the `Content-Type` header must be set to `application/json` for this, as the request may otherwise fail with error 400.\nWhen clients have a valid access token, e.g. obtained through the `POST /api/v1/session` or `POST /api/v1/oauth/token` endpoint, they can use a standard Bearer Authorization header to authenticate their requests. Submitting the access token with a custom `X-Auth-Token` header is supported as well.",
        "title": "PhotoPrism API",
        "contact": {},
        "version": "v1"
    },
    "host": "demo.photoprism.app",
    "paths": {
...
        "/api/v1/photos/{uid}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Photos"
                ],
                "summary": "returns picture details as JSON",
                "operationId": "GetPhoto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Photo UID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Photo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Photos"
                ],
                "summary": "updates picture details and returns them as JSON",
                "operationId": "UpdatePhoto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Photo UID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "properties to be updated (only submit values that should be changed)",
                        "name": "photo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.Photo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Photo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/photos/{uid}/approve": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Photos"
                ],
                "summary": "marks a photo in review as approved",
                "operationId": "ApprovePhoto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "photo uid",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/photos/{uid}/dl": {
            "get": {
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Images",
                    "Files"
                ],
                "summary": "returns the primary file matching that belongs to the photo",
                "operationId": "GetPhotoDownload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "photo uid",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
...
        "/api/v1/photos/{uid}/files/{fileuid}/primary": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Photos",
                    "Stacks"
                ],
                "summary": "sets the primary file for a photo",
                "operationId": "PhotoPrimary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "photo uid",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "file uid",
                        "name": "fileuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Photo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/photos/{uid}/yaml": {
            "get": {
                "produces": [
                    "text/x-yaml"
                ],
                "tags": [
                    "Photos"
                ],
                "summary": "returns picture details as YAML",
                "operationId": "GetPhotoYaml",
                "parameters": [
                    {
                        "type": "string",
                        "description": "photo uid",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Album": {
            "type": "object",
            "properties": {
                "Caption": {
                    "type": "string"
                },
                "Category": {
                    "type": "string"
                },
                "Country": {
                    "type": "string"
                },
                "CreatedAt": {
                    "type": "string"
                },
                "CreatedBy": {
                    "type": "string"
                },
                "Day": {
                    "type": "integer"
                },
                "DeletedAt": {
                    "type": "string"
                },
                "Description": {
                    "type": "string"
                },
                "Favorite": {
                    "type": "boolean"
                },
                "Filter": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "Location": {
                    "type": "string"
                },
                "Month": {
                    "type": "integer"
                },
                "Notes": {
                    "type": "string"
                },
                "Order": {
                    "type": "string"
                },
                "ParentUID": {
                    "type": "string"
                },
                "Path": {
                    "type": "string"
                },
                "Private": {
                    "type": "boolean"
                },
                "PublishedAt": {
                    "type": "string"
                },
                "Slug": {
                    "type": "string"
                },
                "State": {
                    "type": "string"
                },
                "Template": {
                    "type": "string"
                },
                "Thumb": {
                    "type": "string"
                },
                "ThumbSrc": {
                    "type": "string"
                },
                "Title": {
                    "type": "string"
                },
                "Type": {
                    "type": "string"
                },
                "UID": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                },
                "Year": {
                    "type": "integer"
                }
            }
        },
        "entity.Camera": {
            "type": "object",
            "properties": {
                "Description": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "Make": {
                    "type": "string"
                },
                "Model": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                },
                "Notes": {
                    "type": "string"
                },
                "Slug": {
                    "type": "string"
                },
                "Type": {
                    "type": "string"
                }
            }
        },
        "entity.Cell": {
            "type": "object",
            "properties": {
                "Category": {
                    "type": "string"
                },
                "CreatedAt": {
                    "type": "string"
                },
                "ID": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                },
                "Place": {
                    "$ref": "#/definitions/entity.Place"
                },
                "Postcode": {
                    "type": "string"
                },
                "Street": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                }
            }
        },
...
        "entity.Lens": {
            "type": "object",
            "properties": {
                "Description": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "Make": {
                    "type": "string"
                },
                "Model": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                },
                "Notes": {
                    "type": "string"
                },
                "Slug": {
                    "type": "string"
                },
                "Type": {
                    "type": "string"
                }
            }
        },
...
        "entity.Photo": {
            "type": "object",
            "properties": {
                "Albums": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Album"
                    }
                },
                "Altitude": {
                    "type": "integer"
                },
                "Camera": {
                    "$ref": "#/definitions/entity.Camera"
                },
                "CameraID": {
                    "type": "integer"
                },
                "CameraSerial": {
                    "type": "string"
                },
                "CameraSrc": {
                    "type": "string"
                },
                "Caption": {
                    "type": "string"
                },
                "CaptionSrc": {
                    "type": "string"
                },
                "Cell": {
                    "$ref": "#/definitions/entity.Cell"
                },
                "CellAccuracy": {
                    "type": "integer"
                },
                "CellID": {
                    "type": "string"
                },
                "Color": {
                    "type": "integer"
                },
                "Country": {
                    "type": "string"
                },
                "CreatedBy": {
                    "type": "string"
                },
                "Day": {
                    "type": "integer"
                },
                "Details": {
                    "$ref": "#/definitions/entity.Details"
                },
                "DocumentID": {
                    "type": "string"
                },
                "Duration": {
                    "$ref": "#/definitions/time.Duration"
                },
                "EstimatedAt": {
                    "type": "string"
                },
                "Exposure": {
                    "type": "string"
                },
                "FNumber": {
                    "type": "number"
                },
                "Faces": {
                    "type": "integer"
                },
                "Favorite": {
                    "type": "boolean"
                },
                "FocalLength": {
                    "type": "integer"
                },
                "Iso": {
                    "type": "integer"
                },
                "Lat": {
                    "type": "number"
                },
                "Lens": {
                    "$ref": "#/definitions/entity.Lens"
                },
                "LensID": {
                    "type": "integer"
                },
                "Lng": {
                    "type": "number"
                },
                "Month": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                },
                "OriginalName": {
                    "type": "string"
                },
                "Panorama": {
                    "type": "boolean"
                },
                "Path": {
                    "type": "string"
                },
                "Place": {
                    "$ref": "#/definitions/entity.Place"
                },
                "PlaceID": {
                    "type": "string"
                },
                "PlaceSrc": {
                    "type": "string"
                },
                "Private": {
                    "type": "boolean"
                },
                "PublishedAt": {
                    "type": "string"
                },
                "Quality": {
                    "type": "integer"
                },
                "Resolution": {
                    "type": "integer"
                },
                "Scan": {
                    "type": "boolean"
                },
                "Stack": {
                    "type": "integer"
                },
                "TakenAt": {
                    "type": "string"
                },
                "TakenAtLocal": {
                    "type": "string"
                },
                "TakenSrc": {
                    "type": "string"
                },
                "TimeZone": {
                    "type": "string"
                },
                "Title": {
                    "type": "string"
                },
                "TitleSrc": {
                    "type": "string"
                },
                "Type": {
                    "type": "string"
                },
                "TypeSrc": {
                    "type": "string"
                },
                "UID": {
                    "type": "string"
                },
                "Year": {
                    "type": "integer"
                },
                "checkedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "editedAt": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.File"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.PhotoLabel"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "form.Photo": {
            "type": "object",
            "properties": {
                "Altitude": {
                    "type": "integer"
                },
                "CameraID": {
                    "type": "integer"
                },
                "CameraSrc": {
                    "type": "string"
                },
                "Caption": {
                    "type": "string"
                },
                "CaptionSrc": {
                    "type": "string"
                },
                "CellAccuracy": {
                    "type": "integer"
                },
                "CellID": {
                    "type": "string"
                },
                "Country": {
                    "type": "string"
                },
                "Day": {
                    "type": "integer"
                },
                "Details": {
                    "$ref": "#/definitions/form.Details"
                },
                "Exposure": {
                    "type": "string"
                },
                "FNumber": {
                    "type": "number"
                },
                "Favorite": {
                    "type": "boolean"
                },
                "FocalLength": {
                    "type": "integer"
                },
                "Iso": {
                    "type": "integer"
                },
                "Lat": {
                    "type": "number"
                },
                "LensID": {
                    "type": "integer"
                },
                "Lng": {
                    "type": "number"
                },
                "Month": {
                    "type": "integer"
                },
                "OriginalName": {
                    "type": "string"
                },
                "Panorama": {
                    "type": "boolean"
                },
                "PlaceID": {
                    "type": "string"
                },
                "PlaceSrc": {
                    "type": "string"
                },
                "Private": {
                    "type": "boolean"
                },
                "Scan": {
                    "type": "boolean"
                },
                "Stack": {
                    "type": "integer"
                },
                "TakenAt": {
                    "type": "string"
                },
                "TakenAtLocal": {
                    "type": "string"
                },
                "TakenSrc": {
                    "type": "string"
                },
                "TimeZone": {
                    "type": "string"
                },
                "Title": {
                    "type": "string"
                },
                "TitleSrc": {
                    "type": "string"
                },
                "Type": {
                    "type": "string"
                },
                "TypeSrc": {
                    "type": "string"
                },
                "Year": {
                    "type": "integer"
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
        "i18n.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
...
}